Move core settings-related files inside features/settings directory
